requirements.txt
asgiref==3.3.4
astroid==2.12.12
attrs==22.1.0
black==22.10.0
certifi
charset-normalizer==2.1.1
chart-studio==1.1.0
click==8.0.1
colorama==0.4.4
contourpy==1.0.5
cycler==0.11.0
dill==0.3.6
exceptiongroup==1.0.0rc9
fastapi==0.65.2
fonttools==4.38.0
greenlet==1.1.0
h11==0.12.0
idna==3.4
imageio==2.22.2
iniconfig==1.1.1
isort==5.10.1
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
matplotlib==3.6.1
mccabe==0.7.0
mypy-extensions==0.4.3
networkx==2.8.7
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
psycopg2-binary==2.9.1
pydantic==1.8.2
pylint==2.15.5
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyWavelets==1.4.1
requests==2.28.1
retrying==1.3.3
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.19
starlette==0.14.2
tenacity==8.1.0
tifffile==2022.10.10
tomli==2.0.1
tomlkit==0.11.5
typing-extensions==3.10.0.0
urllib3==1.26.12
uvicorn==0.14.0
wrapt==1.14.1
